+++ programs/clearing_house/src/lib.rs
@@ -333,7 +333,7 @@ pub mod clearing_house {
 
     pub fn initialize_serum_fulfillment_config(
         ctx: Context<InitializeSerumFulfillmentConfig>,
-        market_index: u64,
+        market_index: u16,
     ) -> Result<()> {
         validate!(
             market_index != 0,
@@ -638,7 +638,7 @@ pub mod clearing_house {
 
     pub fn deposit(
         ctx: Context<Deposit>,
-        market_index: u64,
+        market_index: u16,
         amount: u64,
         reduce_only: bool,
     ) -> Result<()> {
@@ -738,7 +738,7 @@ pub mod clearing_house {
     )]
     pub fn withdraw(
         ctx: Context<Withdraw>,
-        market_index: u64,
+        market_index: u16,
         amount: u64,
         reduce_only: bool,
     ) -> Result<()> {
@@ -842,7 +842,7 @@ pub mod clearing_house {
 
     pub fn transfer_deposit(
         ctx: Context<TransferDeposit>,
-        market_index: u64,
+        market_index: u16,
         amount: u64,
     ) -> Result<()> {
         let authority_key = ctx.accounts.authority.key;
@@ -996,7 +996,7 @@ pub mod clearing_house {
     #[access_control(
         exchange_not_paused(&ctx.accounts.state)
     )]
-    pub fn settle_lp<'info>(ctx: Context<SettleLP>, market_index: u64) -> Result<()> {
+    pub fn settle_lp<'info>(ctx: Context<SettleLP>, market_index: u16) -> Result<()> {
         let user_key = ctx.accounts.user.key();
         let user = &mut load_mut!(ctx.accounts.user)?;
 
@@ -1029,7 +1029,7 @@ pub mod clearing_house {
     pub fn remove_liquidity<'info>(
         ctx: Context<AddRemoveLiquidity>,
         shares_to_burn: u128,
-        market_index: u64,
+        market_index: u16,
     ) -> Result<()> {
         let user_key = ctx.accounts.user.key();
         let user = &mut load_mut!(ctx.accounts.user)?;
@@ -1108,7 +1108,7 @@ pub mod clearing_house {
     pub fn add_liquidity<'info>(
         ctx: Context<AddRemoveLiquidity>,
         n_shares: u128,
-        market_index: u64,
+        market_index: u16,
     ) -> Result<()> {
         let user_key = ctx.accounts.user.key();
         let user = &mut load_mut!(ctx.accounts.user)?;
@@ -1677,7 +1677,7 @@ pub mod clearing_house {
     #[access_control(
         exchange_not_paused(&ctx.accounts.state)
     )]
-    pub fn update_amms(ctx: Context<UpdateAMM>, market_indexes: [u64; 5]) -> Result<()> {
+    pub fn update_amms(ctx: Context<UpdateAMM>, market_indexes: [u16; 5]) -> Result<()> {
         // up to ~60k compute units (per amm) worst case
 
         let clock = Clock::get()?;
@@ -1699,7 +1699,7 @@ pub mod clearing_house {
     #[access_control(
         exchange_not_paused(&ctx.accounts.state)
     )]
-    pub fn settle_pnl(ctx: Context<SettlePNL>, market_index: u64) -> Result<()> {
+    pub fn settle_pnl(ctx: Context<SettlePNL>, market_index: u16) -> Result<()> {
         let clock = Clock::get()?;
         let state = &ctx.accounts.state;
 
@@ -1745,7 +1745,7 @@ pub mod clearing_house {
     #[access_control(
         exchange_not_paused(&ctx.accounts.state)
     )]
-    pub fn settle_expired_market(ctx: Context<UpdateAMM>, market_index: u64) -> Result<()> {
+    pub fn settle_expired_market(ctx: Context<UpdateAMM>, market_index: u16) -> Result<()> {
         let clock = Clock::get()?;
         let _now = clock.unix_timestamp;
         let state = &ctx.accounts.state;
@@ -1780,7 +1780,7 @@ pub mod clearing_house {
     #[access_control(
         exchange_not_paused(&ctx.accounts.state)
     )]
-    pub fn settle_expired_position(ctx: Context<SettlePNL>, market_index: u64) -> Result<()> {
+    pub fn settle_expired_position(ctx: Context<SettlePNL>, market_index: u16) -> Result<()> {
         let clock = Clock::get()?;
         let state = &ctx.accounts.state;
 
@@ -1831,7 +1831,7 @@ pub mod clearing_house {
     )]
     pub fn liquidate_perp(
         ctx: Context<LiquidatePerp>,
-        market_index: u64,
+        market_index: u16,
         liquidator_max_base_asset_amount: u128,
     ) -> Result<()> {
         let clock = Clock::get()?;
@@ -1887,8 +1887,8 @@ pub mod clearing_house {
     )]
     pub fn liquidate_borrow(
         ctx: Context<LiquidateBorrow>,
-        asset_market_index: u64,
-        liability_market_index: u64,
+        asset_market_index: u16,
+        liability_market_index: u16,
         liquidator_max_liability_transfer: u128,
     ) -> Result<()> {
         let clock = Clock::get()?;
@@ -1943,8 +1943,8 @@ pub mod clearing_house {
     )]
     pub fn liquidate_borrow_for_perp_pnl(
         ctx: Context<LiquidateBorrowForPerpPnl>,
-        perp_market_index: u64,
-        spot_market_index: u64,
+        perp_market_index: u16,
+        spot_market_index: u16,
         liquidator_max_liability_transfer: u128,
     ) -> Result<()> {
         let clock = Clock::get()?;
@@ -1998,8 +1998,8 @@ pub mod clearing_house {
     )]
     pub fn liquidate_perp_pnl_for_deposit(
         ctx: Context<LiquidatePerpPnlForDeposit>,
-        perp_market_index: u64,
-        spot_market_index: u64,
+        perp_market_index: u16,
+        spot_market_index: u16,
         liquidator_max_pnl_transfer: u128,
     ) -> Result<()> {
         let state = &ctx.accounts.state;
@@ -2050,8 +2050,8 @@ pub mod clearing_house {
 
     pub fn resolve_perp_pnl_deficit(
         ctx: Context<ResolvePerpPnlDeficit>,
-        spot_market_index: u64,
-        perp_market_index: u64,
+        spot_market_index: u16,
+        perp_market_index: u16,
     ) -> Result<()> {
         let clock = Clock::get()?;
         let now = clock.unix_timestamp;
@@ -2153,8 +2153,8 @@ pub mod clearing_house {
     )]
     pub fn resolve_perp_bankruptcy(
         ctx: Context<ResolveBankruptcy>,
-        quote_spot_market_index: u64,
-        market_index: u64,
+        quote_spot_market_index: u16,
+        market_index: u16,
     ) -> Result<()> {
         let clock = Clock::get()?;
         let now = clock.unix_timestamp;
@@ -2235,7 +2235,7 @@ pub mod clearing_house {
     )]
     pub fn resolve_borrow_bankruptcy(
         ctx: Context<ResolveBankruptcy>,
-        market_index: u64,
+        market_index: u16,
     ) -> Result<()> {
         let state = &ctx.accounts.state;
         let clock = Clock::get()?;
@@ -2779,7 +2779,7 @@ pub mod clearing_house {
         exchange_not_paused(&ctx.accounts.state) &&
         valid_oracle_for_market(&ctx.accounts.oracle, &ctx.accounts.market)
     )]
-    pub fn update_funding_rate(ctx: Context<UpdateFundingRate>, market_index: u64) -> Result<()> {
+    pub fn update_funding_rate(ctx: Context<UpdateFundingRate>, market_index: u16) -> Result<()> {
         let market = &mut load_mut!(ctx.accounts.market)?;
         let clock = Clock::get()?;
         let now = clock.unix_timestamp;
@@ -3120,7 +3120,7 @@ pub mod clearing_house {
 
     pub fn update_spot_market_if_factor(
         ctx: Context<AdminUpdateSpotMarket>,
-        spot_market_index: u64,
+        spot_market_index: u16,
         user_if_factor: u32,
         total_if_factor: u32,
     ) -> Result<()> {
@@ -3497,7 +3497,7 @@ pub mod clearing_house {
 
     pub fn initialize_insurance_fund_stake(
         ctx: Context<InitializeInsuranceFundStake>,
-        market_index: u64,
+        market_index: u16,
     ) -> Result<()> {
         let mut if_stake = ctx
             .accounts
@@ -3515,7 +3515,7 @@ pub mod clearing_house {
 
     pub fn settle_revenue_to_insurance_fund(
         ctx: Context<SettleRevenueToInsuranceFund>,
-        _market_index: u64,
+        _market_index: u16,
     ) -> Result<()> {
         let state = &ctx.accounts.state;
         let spot_market = &mut load_mut!(ctx.accounts.spot_market)?;
@@ -3571,7 +3571,7 @@ pub mod clearing_house {
 
     pub fn add_insurance_fund_stake(
         ctx: Context<AddInsuranceFundStake>,
-        market_index: u64,
+        market_index: u16,
         amount: u64,
     ) -> Result<()> {
         if amount == 0 {
@@ -3618,7 +3618,7 @@ pub mod clearing_house {
 
     pub fn request_remove_insurance_fund_stake(
         ctx: Context<RequestRemoveInsuranceFundStake>,
-        market_index: u64,
+        market_index: u16,
         amount: u64,
     ) -> Result<()> {
         let clock = Clock::get()?;
@@ -3667,7 +3667,7 @@ pub mod clearing_house {
 
     pub fn cancel_request_remove_insurance_fund_stake(
         ctx: Context<RequestRemoveInsuranceFundStake>,
-        market_index: u64,
+        market_index: u16,
     ) -> Result<()> {
         let clock = Clock::get()?;
         let now = clock.unix_timestamp;
@@ -3700,7 +3700,7 @@ pub mod clearing_house {
 
     pub fn remove_insurance_fund_stake(
         ctx: Context<RemoveInsuranceFundStake>,
-        market_index: u64,
+        market_index: u16,
     ) -> Result<()> {
         let clock = Clock::get()?;
         let now = clock.unix_timestamp;
@@ -3743,7 +3743,7 @@ pub mod clearing_house {
 
     pub fn admin_remove_insurance_fund_stake(
         ctx: Context<AdminRemoveInsuranceFundStake>,
-        market_index: u64,
+        market_index: u16,
         amount: u64,
     ) -> Result<()> {
         let clock = Clock::get()?;
