+++ programs/clearing_house/src/state/events.rs
@@ -25,7 +25,7 @@ pub struct DepositRecord {
     pub user: Pubkey,
     pub direction: DepositDirection,
     pub amount: u64,
-    pub market_index: u64,
+    pub market_index: u16,
     pub oracle_price: i128,
     pub referrer: Pubkey,
     pub from: Option<Pubkey>,
@@ -50,7 +50,7 @@ pub struct FundingPaymentRecord {
     pub ts: i64,
     pub user_authority: Pubkey,
     pub user: Pubkey,
-    pub market_index: u64,
+    pub market_index: u16,
     pub funding_payment: i128,
     pub base_asset_amount: i128,
     pub user_last_cumulative_funding: i128,
@@ -63,7 +63,7 @@ pub struct FundingPaymentRecord {
 pub struct FundingRateRecord {
     pub ts: i64,
     pub record_id: u64,
-    pub market_index: u64,
+    pub market_index: u16,
     pub funding_rate: i128,
     pub funding_rate_long: i128,
     pub funding_rate_short: i128,
@@ -80,7 +80,7 @@ pub struct FundingRateRecord {
 pub struct CurveRecord {
     pub ts: i64,
     pub record_id: u64,
-    pub market_index: u64,
+    pub market_index: u16,
     pub peg_multiplier_before: u128,
     pub base_asset_reserve_before: u128,
     pub quote_asset_reserve_before: u128,
@@ -112,7 +112,7 @@ pub struct OrderActionRecord {
     pub ts: i64,
     pub action: OrderAction,
     pub action_explanation: OrderActionExplanation,
-    pub market_index: u64,
+    pub market_index: u16,
     pub market_type: MarketType,
 
     pub filler: Option<Pubkey>,
@@ -155,7 +155,7 @@ pub fn get_order_action_record(
     ts: i64,
     action: OrderAction,
     action_explanation: OrderActionExplanation,
-    market_index: u64,
+    market_index: u16,
     filler: Option<Pubkey>,
     fill_record_id: Option<u64>,
     filler_reward: Option<u128>,
@@ -296,7 +296,7 @@ pub struct LPRecord {
     pub user: Pubkey,
     pub action: LPAction,
     pub n_shares: u128,
-    pub market_index: u64,
+    pub market_index: u16,
     pub delta_base_asset_amount: i128,
     pub delta_quote_asset_amount: i128,
     pub pnl: i128,
@@ -354,7 +354,7 @@ impl Default for LiquidationType {
 
 #[derive(AnchorSerialize, AnchorDeserialize, Clone, Debug, Default)]
 pub struct LiquidatePerpRecord {
-    pub market_index: u64,
+    pub market_index: u16,
     pub oracle_price: i128,
     pub base_asset_amount: i128,
     pub quote_asset_amount: i128,
@@ -369,10 +369,10 @@ pub struct LiquidatePerpRecord {
 
 #[derive(AnchorSerialize, AnchorDeserialize, Clone, Debug, Default)]
 pub struct LiquidateBorrowRecord {
-    pub asset_market_index: u64,
+    pub asset_market_index: u16,
     pub asset_price: i128,
     pub asset_transfer: u128,
-    pub liability_market_index: u64,
+    pub liability_market_index: u16,
     pub liability_price: i128,
     pub liability_transfer: u128,
     pub if_fee: u64,
@@ -380,27 +380,27 @@ pub struct LiquidateBorrowRecord {
 
 #[derive(AnchorSerialize, AnchorDeserialize, Clone, Debug, Default)]
 pub struct LiquidateBorrowForPerpPnlRecord {
-    pub perp_market_index: u64,
+    pub perp_market_index: u16,
     pub market_oracle_price: i128,
     pub pnl_transfer: u128,
-    pub liability_market_index: u64,
+    pub liability_market_index: u16,
     pub liability_price: i128,
     pub liability_transfer: u128,
 }
 
 #[derive(AnchorSerialize, AnchorDeserialize, Clone, Debug, Default)]
 pub struct LiquidatePerpPnlForDepositRecord {
-    pub perp_market_index: u64,
+    pub perp_market_index: u16,
     pub market_oracle_price: i128,
     pub pnl_transfer: u128,
-    pub asset_market_index: u64,
+    pub asset_market_index: u16,
     pub asset_price: i128,
     pub asset_transfer: u128,
 }
 
 #[derive(AnchorSerialize, AnchorDeserialize, Clone, Debug, Default)]
 pub struct PerpBankruptcyRecord {
-    pub market_index: u64,
+    pub market_index: u16,
     pub pnl: i128,
     pub if_payment: u128,
     pub cumulative_funding_rate_delta: i128,
@@ -408,7 +408,7 @@ pub struct PerpBankruptcyRecord {
 
 #[derive(AnchorSerialize, AnchorDeserialize, Clone, Debug, Default)]
 pub struct BorrowBankruptcyRecord {
-    pub market_index: u64,
+    pub market_index: u16,
     pub borrow_amount: u128,
     pub if_payment: u128,
     pub cumulative_deposit_interest_delta: u128,
@@ -419,7 +419,7 @@ pub struct BorrowBankruptcyRecord {
 pub struct SettlePnlRecord {
     pub ts: i64,
     pub user: Pubkey,
-    pub market_index: u64,
+    pub market_index: u16,
     pub pnl: i128,
     pub base_asset_amount: i128,
     pub quote_asset_amount_after: i128,
@@ -431,8 +431,8 @@ pub struct SettlePnlRecord {
 #[derive(Default)]
 pub struct InsuranceFundRecord {
     pub ts: i64,
-    pub spot_market_index: u64,
-    pub perp_market_index: u64,
+    pub spot_market_index: u16,
+    pub perp_market_index: u16,
     pub user_if_factor: u32,
     pub total_if_factor: u32,
     pub vault_amount_before: u64,
@@ -449,7 +449,7 @@ pub struct InsuranceFundStakeRecord {
     pub user_authority: Pubkey,
     pub action: StakeAction,
     pub amount: u64,
-    pub market_index: u64,
+    pub market_index: u16,
 
     pub insurance_vault_amount_before: u64,
     pub if_shares_before: u128,
